# packages/server/dist/swagger.json
{
	"basePath": "/",
	"definitions": {
		"AccountResponse": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"username": {
					"type": "string",
					"description": ""
				},
				"nick": {
					"type": "string",
					"description": ""
				},
				"avatar": {
					"type": "string",
					"description": ""
				},
				"type": {
					"type": "string",
					"description": ""
				},
				"groups": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"roles": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"email": {
					"type": "string",
					"description": ""
				},
				"mobile": {
					"type": "string",
					"description": ""
				},
				"isDisable": {
					"type": "boolean",
					"description": ""
				},
				"isAdmin": {
					"type": "boolean",
					"description": ""
				},
				"isApproved": {
					"type": "boolean",
					"description": ""
				},
				"expired": {
					"type": "string",
					"format": "date-time",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"username",
				"nick",
				"avatar",
				"type",
				"groups",
				"roles",
				"email",
				"mobile",
				"isDisable",
				"isAdmin",
				"isApproved",
				"expired"
			]
		},
		"CreateAccountDto": {
			"description": "",
			"properties": {
				"avatar": {
					"type": "string",
					"description": ""
				},
				"email": {
					"type": "string",
					"description": ""
				},
				"mobile": {
					"type": "string",
					"description": ""
				},
				"groups": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"roles": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"username": {
					"type": "string",
					"description": ""
				},
				"password": {
					"type": "string",
					"description": ""
				},
				"nick": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"username",
				"password",
				"nick"
			]
		},
		"EditAccountDto": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"avatar": {
					"type": "string",
					"description": ""
				},
				"email": {
					"type": "string",
					"description": ""
				},
				"groups": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"roles": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"mobile": {
					"type": "string",
					"description": ""
				},
				"password": {
					"type": "string",
					"description": ""
				},
				"nick": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"nick"
			]
		},
		"Function": {
			"description": "Creates a new function.",
			"properties": {
				"prototype": {
					"type": "object",
					"description": ""
				},
				"length": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"arguments": {
					"type": "object",
					"description": ""
				},
				"caller": {
					"$ref": "#/definitions/Function"
				}
			},
			"type": "object",
			"required": [
				"prototype",
				"length",
				"arguments",
				"caller"
			]
		},
...
		"KeyValue": {
			"description": "",
			"properties": {
				"label": {
					"type": "string",
					"description": ""
				},
				"value": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"label",
				"value"
			]
		},
		"Error": {
			"description": "",
			"properties": {
				"name": {
					"type": "string",
					"description": ""
				},
				"message": {
					"type": "string",
					"description": ""
				},
				"stack": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"message"
			]
		},
		"PaginateAccount": {
			"description": "",
			"properties": {
				"error": {
					"$ref": "#/definitions/Error"
				},
				"list": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/AccountResponse"
					},
					"description": ""
				},
				"total": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"list",
				"total"
			]
		},
...
		"Object": {
			"description": "Provides functionality common to all JavaScript objects.",
			"properties": {
				"constructor": {
					"$ref": "#/definitions/Function"
				}
			},
			"type": "object",
			"required": [
				"constructor"
			]
		},
		"DictResponse": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"category": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"translate": {
					"type": "string",
					"description": ""
				},
				"expand": {
					"$ref": "#/definitions/Object"
				}
			},
			"type": "object",
			"required": [
				"id",
				"category",
				"name",
				"translate",
				"expand"
			]
		},
		"CreateDictDto": {
			"description": "",
			"properties": {
				"category": {
					"type": "string",
					"description": ""
				},
				"translate": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"expand": {
					"$ref": "#/definitions/Object"
				}
			},
			"type": "object",
			"required": [
				"category",
				"translate",
				"name",
				"expand"
			]
		},
		"EditDictDto": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"category": {
					"type": "string",
					"description": ""
				},
				"translate": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"expand": {
					"$ref": "#/definitions/Object"
				}
			},
			"type": "object",
			"required": [
				"id",
				"category",
				"translate",
				"name",
				"expand"
			]
		},
		"PaginateDict": {
			"description": "",
			"properties": {
				"error": {
					"$ref": "#/definitions/Error"
				},
				"list": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/DictResponse"
					},
					"description": ""
				},
				"total": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"list",
				"total"
			]
		},
...
		"LogResponse": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"operator": {
					"type": "string",
					"description": ""
				},
				"operatorIp": {
					"type": "string",
					"description": ""
				},
				"operation": {
					"type": "string",
					"description": ""
				},
				"comment": {
					"type": "string",
					"description": ""
				},
				"createdAt": {
					"type": "string",
					"format": "date-time",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"operator",
				"operatorIp",
				"operation",
				"comment",
				"createdAt"
			]
		},
		"PaginateLog": {
			"description": "",
			"properties": {
				"error": {
					"$ref": "#/definitions/Error"
				},
				"list": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/LogResponse"
					},
					"description": ""
				},
				"total": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"list",
				"total"
			]
		},
		"GroupResponse": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"outid": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"icon": {
					"type": "string",
					"description": ""
				},
				"parent": {
					"type": "string",
					"description": ""
				},
				"paths": {
					"type": "array",
					"items": {
						"type": "object"
					},
					"description": ""
				},
				"director": {
					"type": "string",
					"description": ""
				},
				"order": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"isRegion": {
					"type": "boolean",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"order"
			]
		},
		"CreateGroupDto": {
			"description": "",
			"properties": {
				"outid": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"icon": {
					"type": "string",
					"description": ""
				},
				"parent": {
					"type": "string",
					"description": ""
				},
				"paths": {
					"type": "array",
					"items": {
						"type": "object"
					},
					"description": ""
				},
				"director": {
					"type": "string",
					"description": ""
				},
				"order": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"isRegion": {
					"type": "boolean",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"order"
			]
		},
		"EditGroupDto": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"outid": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"icon": {
					"type": "string",
					"description": ""
				},
				"parent": {
					"type": "string",
					"description": ""
				},
				"paths": {
					"type": "array",
					"items": {
						"type": "object"
					},
					"description": ""
				},
				"director": {
					"type": "string",
					"description": ""
				},
				"order": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"isRegion": {
					"type": "boolean",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"order"
			]
		},
		"PaginateGroup": {
			"description": "",
			"properties": {
				"error": {
					"$ref": "#/definitions/Error"
				},
				"list": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/GroupResponse"
					},
					"description": ""
				},
				"total": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"list",
				"total"
			]
		}
	},
	"info": {
		"description": "项目标准接口",
		"license": {
			"name": "MIT"
		},
		"title": "rest api interface",
		"version": "0.0.1"
	},
	"paths": {
		"/api/account": {
			"post": {
				"operationId": "AccountCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AccountResponse"
						}
					}
				},
				"description": "创建帐号",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "帐号信息",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateAccountDto"
						}
					}
				]
			},
			"put": {
				"operationId": "AccountUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AccountResponse"
						}
					}
				},
				"description": "更新帐号",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "帐号信息",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/EditAccountDto"
						}
					}
				]
			}
		},
...
		"/api/account/search": {
			"get": {
				"operationId": "AccountGetAccountsByKeyword",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/KeyValue"
							}
						}
					}
				},
				"description": "按关键词查询帐号",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "value",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/account/query": {
			"get": {
				"operationId": "AccountQuery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginateAccount"
						}
					}
				},
				"description": "分页查询帐号数据",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "group",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "role",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "page",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "",
						"in": "query",
						"name": "size",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "",
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
...
		"/api/account/{id}": {
			"delete": {
				"operationId": "AccountRemove",
				"produces": [
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"description": "删除帐号",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "帐号编号",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"get": {
				"operationId": "AccountGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AccountResponse"
						}
					}
				},
				"description": "查询帐号",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "编号",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/api/dict/search": {
			"get": {
				"operationId": "DictSearch",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/KeyValue"
							}
						}
					}
				},
				"description": "搜索字典表",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "键",
						"in": "query",
						"name": "value",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "category",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/dict": {
			"post": {
				"operationId": "DictCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/DictResponse"
						}
					}
				},
				"description": "创建字典",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "设置项实体",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateDictDto"
						}
					}
				]
			},
			"put": {
				"operationId": "DictUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/DictResponse"
						}
					}
				},
				"description": "更新字典",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "设置项实体",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/EditDictDto"
						}
					}
				]
			}
		},
		"/api/dict/query": {
			"get": {
				"operationId": "DictQuery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginateDict"
						}
					}
				},
				"description": "分页查询字典",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "category",
						"required": false,
						"type": "string"
					},
					{
						"description": "第几页",
						"in": "query",
						"name": "page",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "页大小",
						"in": "query",
						"name": "size",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "排序",
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/dict/{id}": {
			"get": {
				"operationId": "DictGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/DictResponse"
						}
					}
				},
				"description": "按编号获取字典",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "DictRemove",
				"produces": [
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"description": "删除字典",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/api/log/search": {
			"get": {
				"operationId": "LogSearch",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/KeyValue"
							}
						}
					}
				},
				"description": "搜索日志",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "键",
						"in": "query",
						"name": "value",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/log/query": {
			"get": {
				"operationId": "LogQuery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginateLog"
						}
					}
				},
				"description": "分页查询日志",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "第几页",
						"in": "query",
						"name": "page",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "页大小",
						"in": "query",
						"name": "size",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "排序",
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/log/{id}": {
			"get": {
				"operationId": "LogGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/LogResponse"
						}
					}
				},
				"description": "按编号获取日志",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/api/group/search": {
			"get": {
				"operationId": "GroupSearch",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/KeyValue"
							}
						}
					}
				},
				"description": "查询用户组",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "已选中的用户组编号",
						"in": "query",
						"name": "value",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/group": {
			"post": {
				"operationId": "GroupCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/GroupResponse"
						}
					}
				},
				"description": "创建用户组",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "创建参数",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateGroupDto"
						}
					}
				]
			},
			"put": {
				"operationId": "GroupUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/GroupResponse"
						}
					}
				},
				"description": "更新用户组",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "用户组参数",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/EditGroupDto"
						}
					}
				]
			}
		},
		"/api/group/query": {
			"get": {
				"operationId": "GroupQuery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginateGroup"
						}
					}
				},
				"description": "查询用户组数据",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "isGroup",
						"required": false,
						"type": "boolean"
					},
					{
						"description": "第几页",
						"in": "query",
						"name": "page",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "页大小",
						"in": "query",
						"name": "size",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "排序",
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/group/{id}": {
			"delete": {
				"operationId": "GroupRemove",
				"produces": [
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"description": "删除用户组",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "用户组编号",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"get": {
				"operationId": "GroupGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/GroupResponse"
						}
					}
				},
				"description": "查询用户组",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "用户组编号",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"swagger": "2.0",
	"securityDefinitions": {},
	"produces": [
		"application/json"
	]
}